+++ test/dummy/app/frontend/packs/application.js
@@ -1,5 +1,6 @@
 import ReactOnRails from 'react-on-rails';
-// import ReactSample from '../javascripts/test-react.jsx';
+import Modal from '../javascripts/react/Modal/Modal.jsx';
+import AjaxModal from '../javascripts/react/Modal/AjaxModal.jsx';
 
 // Generate application.css
 import '../stylesheets/application.scss';
@@ -7,6 +8,7 @@ import '../stylesheets/application.scss';
 // Import the application file as the contents of this file
 import '../javascripts/application.js';
 
-// ReactOnRails.register({ 
-//   ReactSample,
-// });
+ReactOnRails.register({ 
+  Modal,
+  AjaxModal,
+});